Moved simulaqron.cqc_backend.cqcProtocol.CQCProtocol and simulaqron.cqc_backend.cqcMessageHandler.CQCMessageHandler to CQC repo and some other small things
