+++ CHANGELOG.md
@@ -3,6 +3,40 @@ CHANGELOG
 
 For more details refer to the [documentation](https://softwarequtech.github.io/SimulaQron/html/index.html).
 
+Upcoming
+--------
+
+
+2019-04-08 (v2.2.0)
+-------------------
+ - The classes `simulaqron.cqc_backend.cqcProtocol.CQCProtocol` and `simulaqron.cqc_backend.cqcMessageHandler.CQCMessageHandler` move to `cqc.Protocol.CQCProtocol` and `cqc.MessageHandler.CQCMessageHandler` respectively in the repo https://github.com/SoftwareQuTech/CQC-Python.
+
+
+2019-04-05 (v2.1.2)
+-------------------
+ - Moved new_ent_id up from simulaqronCQCHandler to CQCMessageHandler
+ 
+
+2019-03-29 (v2.1.1)
+-------------------
+
+Tests can now be run from python from the method `simulaqron.tests`.
+
+
+2019-03-28 (v2.1.0)
+-------------------
+These are changes in Develop will be merged to master.
+
+ - The CQC specification and libraries will move to it's own repos.
+
+ - SimulaQron can now be installed using `pip3 install simulaqron`.
+
+ - If SimulaQron is installed using pip, the command in the terminal will be called `simulaqron` and not `SimulaQron`.
+
+ - There is a new command in the which is useful if something crashed when a simulated network was running. Simply to `simulaqron reset`.
+
+ - All settings can be accessed through the command line interface by the commands `simulaqron set` and `simulaqron get`.
+
 
 2019-03-15 (v1.3)
 -----------------
@@ -22,30 +56,3 @@ These are changes in Develop was merged to master.
   New is also the class `simulaqron.network.Network` which makes is easy to start a network within Python, see docs.
   
 - All test are now unittests and can be started as `make tests` or `make tests_full` for a longer test.
-
-2019-03-28 (v2.1.0)
--------------------
-These are changes in Develop will be merged to master.
-
- - The CQC specification and libraries will move to it's own repos.
-
- - SimulaQron can now be installed using `pip3 install simulaqron`.
-
- - If SimulaQron is installed using pip, the command in the terminal will be called `simulaqron` and not `SimulaQron`.
-
- - There is a new command in the which is useful if something crashed when a simulated network was running. Simply to `simulaqron reset`.
-
- - All settings can be accessed through the command line interface by the commands `simulaqron set` and `simulaqron get`.
-
-2019-03-29 (v2.1.1)
--------------------
-
-Tests can now be run from python from the method `simulaqron.tests`.
-
-
-2019-04-05 (v2.1.2)
--------------------
- - Moved new_ent_id up from simulaqronCQCHandler to CQCMessageHandler
-
-Upcoming
---------
